Reuse template_error/1 and fix changeset :action
