+++ lib/beacon/live_admin/components.ex
@@ -1,5 +1,6 @@
 defmodule Beacon.LiveAdmin.Components do
   use Phoenix.Component
+  import Beacon.LiveAdmin.CoreComponents, only: [error: 1]
   import Beacon.LiveAdmin.Cluster, only: [call: 4]
 
   @doc false
@@ -23,6 +24,30 @@ defmodule Beacon.LiveAdmin.Components do
 
     ~H"""
     <%= {:safe, @raw} %>
+    """
+  end
+
+  @doc """
+  Display an error message for template/body fields.
+  """
+  def template_error(field) do
+    {message, compilation_error} =
+      case field.errors do
+        [{message, [compilation_error: compilation_error]} | _] -> {message, compilation_error}
+        [{message, _}] -> {message, nil}
+        _ -> {nil, nil}
+      end
+
+    assigns = %{
+      message: message,
+      compilation_error: compilation_error
+    }
+
+    ~H"""
+    <.error :if={@message}><%= @message %></.error>
+    <code :if={@compilation_error} class="mt-3 text-sm text-rose-600 phx-no-feedback:hidden">
+      <pre><%= @compilation_error %></pre>
+    </code>
     """
   end
 end
