+++ lib/beacon/live_admin/page_builder.ex
@@ -52,6 +52,7 @@ defmodule Beacon.LiveAdmin.PageBuilder do
   defmacro __using__(opts) do
     quote location: :keep, bind_quoted: [opts: opts] do
       use Phoenix.Component
+      import Beacon.LiveAdmin.Components, only: [template_error: 1]
       import Beacon.LiveAdmin.CoreComponents
 
       import Beacon.LiveAdmin.Router,
